Change logging from [YT_*] to [LIBYT_*] to avoid confusion.

+++ src/yt_run_InteractiveMode.cpp
@@ -13,15 +13,17 @@
 // Description :  Enter libyt interactive mode.
 //
 // Note        :  1. Only enter this mode when executed inline functions have errors or flag_file_name
-//                   is detacted.
-//                2. Display inline script execute result success/failed, and show errors if have.
+//                   is detected.
+//                2. Display inline script execute result success/failed, and show errors if had.
 //                3. Enter interactive mode, user will be operating in inline script's name space.
 //                   (1) Python scripting
 //                   (2) libyt command
-//                   (3) Execute charactars should be less than INT_MAX.
+//                   (3) Execute characters should be less than INT_MAX.
 //                4. Let user add and decide what inline function to run in the follow process.
+//                5. Tagged every libyt sent messages, so that when other processes (ex: python) also
+//                   use mpi, the processes will hang and won't mess up the behavior.
 //
-// Parameter   :  const char *flag_file_name : once this file is detacted, it enters interactive mode
+// Parameter   :  const char *flag_file_name : once this file is detected, it enters interactive mode
 //
 // Return      :  YT_SUCCESS or YT_FAIL
 //-------------------------------------------------------------------------------------------------------
@@ -149,7 +151,7 @@ int yt_run_InteractiveMode(const char* flag_file_name) {
                     // broadcast to other ranks, code character num no longer than INT_MAX
                     int temp = (int) strlen(code);
                     MPI_Bcast(&temp, 1, MPI_INT, root, MPI_COMM_WORLD);
-                    MPI_Bcast(code, strlen(code), MPI_CHAR, root, MPI_COMM_WORLD);
+                    MPI_Bcast(code, (int) strlen(code), MPI_CHAR, root, MPI_COMM_WORLD);
 #endif
 
                     // run code
